ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
google-genai==1.30.0
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.0
numpy==1.26.4
scikit-learn==1.5.2
Pillow==10.4.0
opencv-python==4.10.0.84
httpx==0.27.2
requests==2.32.3
aiofiles==24.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
prometheus-client==0.21.0
